requirements.txt
bibtexparser==1.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
feedparser==6.0.11
idna==3.6
pyparsing==3.1.2
python-dotenv==1.0.1
pytz==2024.1
pyzotero==1.5.18
requests==2.31.0
sgmllib3k==1.0.0
urllib3==2.2.1
